* Misc fixups during review
StatM.hpp: Big TODO to make this a nice, general interface even if we
  only do the Linux implementation.  Also a few, more minor TODO's.

CMakeLists.txt: remove useless call that tells it to link with no extra libraries.

stat_m.cpp: make int64_t size_t to remove signedness warning.

+++ src/tests/stat_m.cpp
@@ -27,8 +27,8 @@ using boost::format;
 
 int main() {
     StatM statm;
-    int64_t initial_resident = statm.getStatM().resident;
-    int64_t initial_size = statm.getCached().size;
+    size_t initial_resident = statm.getStatM().resident;
+    size_t initial_size = statm.getCached().size;
     LintelLog::info(format("initially %d/%d") % initial_size % initial_resident);
 
     char * trash = static_cast<char *>(malloc(4096 * 10000));
